weeks/week-17/logs-service/requirements.txt
fastapi==0.110.2
uvicorn==0.30.0
pydantic==2.7.1
grpcio==1.62.2
